Create To Do List, Pygame separate application
The file would not properly transfer over so had to copy and paste the code into here.
This is a to do list application that launches onto a separate application. It asks you for your name, if you want to add anything to the list, and then finally if you want to remove anything from the list. It is all displayed on text boxes, my first time dealing with those.